services/auth/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
asyncpg>=0.29.0
python-multipart>=0.0.9
numpy>=1.24.0
Pillow>=10.0.0
facenet-pytorch>=2.5.3
torch>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
